ActivityPub Post mapping 스키마로 정리한다
